requirements.txt
discord.py
git+https://github.com/developomp/wbscraper.git#egg=wbscraper
firebase_admin
nekos.py
